Extract a Contact class
